some small compile fixes
git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@1229653 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ src/libkstapp/datawizard.cpp
@@ -1149,9 +1149,9 @@ void DataWizard::finished() {
     if (layout_type != CurvePlacement::Custom) {
       layout_type = CurvePlacement::Custom;
       if (_pagePlot->plotTabPlacement() == DataWizardPagePlot::SeparateTabs) {
-        num_columns = sqrt(plotList.size()/2);
+        num_columns = sqrt((double)plotList.size()/2);
       } else {
-        num_columns = sqrt(plotList.size());
+        num_columns = sqrt((double)plotList.size());
       }
     }
   }
