+++ src/libkstapp/viewitem.cpp
@@ -18,7 +18,7 @@
 #include "plotitemmanager.h"
 #include "document.h"
 #include "datacollection.h"
-#include "formatgridhelper.cpp"
+#include "formatgridhelper.h"
 
 #include "layoutboxitem.h"
 
@@ -815,7 +815,7 @@ void ViewItem::createAutoLayout() {
 
 void ViewItem::createCustomLayout() {
   bool ok;
-  int default_cols = qMax(1,int(sqrt(Data::self()->plotList().count())));
+  int default_cols = qMax(1,int(sqrt((double)Data::self()->plotList().count())));
   int columns = QInputDialog::getInteger(view(), tr("Kst"),
                                       tr("Select Number of Columns"),default_cols, 0,
                                       10, 1, &ok);
